# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Interview challenge for building a Payment Gateway - Go version",
        "title": "Payment Gateway Challenge Go",
        "contact": {}
    },
    "host": "localhost:8090",
    "basePath": "/",
    "paths": {
        "/api/payments": {
            "post": {
                "description": "Processes a card payment through the payment gateway",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "payments"
                ],
                "summary": "Create a payment",
                "parameters": [
                    {
                        "description": "Payment request",
                        "name": "payment",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.PostPaymentRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.PostPaymentResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.PostPaymentResponse"
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/payments/{id}": {
            "get": {
                "description": "Retrieves a previously processed payment",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "payments"
                ],
                "summary": "Get payment",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Payment ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.PostPaymentResponse"
                        }
                    },
                    "204": {
                        "description": "No Content",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.PaymentStatus": {
            "type": "string",
            "enum": [
                "Authorized",
                "Declined",
                "Rejected"
            ],
            "x-enum-varnames": [
                "Authorized",
                "Declined",
                "Rejected"
            ]
        },
        "models.PostPaymentRequest": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "integer"
                },
                "card_number": {
                    "type": "string"
                },
                "currency": {
                    "type": "string"
                },
                "cvv": {
                    "type": "string"
                },
                "expiry_month": {
                    "type": "integer"
                },
                "expiry_year": {
                    "type": "integer"
                }
            }
        },
        "models.PostPaymentResponse": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "integer"
                },
                "card_number_last_four": {
                    "type": "string"
                },
                "currency": {
                    "type": "string"
                },
                "expiry_month": {
                    "type": "integer"
                },
                "expiry_year": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "payment_status": {
                    "$ref": "#/definitions/models.PaymentStatus"
                }
            }
        }
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        }
    }
}